Move common stuff to common module
