+++ common/src/main/java/io/streamshub/mcp/common/util/TimeRangeValidator.java
@@ -2,7 +2,7 @@
  * Copyright StreamsHub authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.streamshub.mcp.strimzi.util;
+package io.streamshub.mcp.common.util;
 
 import io.quarkiverse.mcp.server.ToolCallException;
