+++ strimzi-mcp/src/main/java/io/streamshub/mcp/strimzi/service/CompletionService.java
@@ -6,6 +6,7 @@ package io.streamshub.mcp.strimzi.service;
 
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.quarkiverse.mcp.server.CompleteContext;
+import io.streamshub.mcp.common.service.CompletionCache;
 import io.streamshub.mcp.common.service.CompletionHelper;
 import io.streamshub.mcp.strimzi.config.StrimziConstants;
 import io.strimzi.api.kafka.model.kafka.Kafka;
@@ -13,13 +14,9 @@ import io.strimzi.api.kafka.model.nodepool.KafkaNodePool;
 import io.strimzi.api.kafka.model.topic.KafkaTopic;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
-import org.eclipse.microprofile.config.inject.ConfigProperty;
-import org.jboss.logging.Logger;
 
-import java.time.Instant;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * Strimzi-specific auto-completion service for MCP prompt and resource template parameters.
@@ -31,7 +28,6 @@ import java.util.concurrent.ConcurrentHashMap;
 @ApplicationScoped
 public class CompletionService {
 
-    private static final Logger LOG = Logger.getLogger(CompletionService.class);
     // Prompt arguments use descriptive names (e.g., "cluster_name")
     private static final String ARG_NAMESPACE = "namespace";
     private static final String ARG_CLUSTER_NAME = "cluster_name";
@@ -43,20 +39,12 @@ public class CompletionService {
     CompletionHelper completionHelper;
 
     @Inject
+    CompletionCache completionCache;
+
+    @Inject
     KubernetesClient kubernetesClient;
 
-    @ConfigProperty(name = "mcp.completion.cache-ttl-seconds", defaultValue = "5")
-    long cacheTtlSeconds;
-
-    private final Map<String, CachedNames> nameCache = new ConcurrentHashMap<>();
-
     CompletionService() {
-    }
-
-    private record CachedNames(List<String> names, Instant expiry) {
-        boolean isExpired() {
-            return Instant.now().isAfter(expiry);
-        }
     }
 
     /**
@@ -154,74 +142,50 @@ public class CompletionService {
         return List.of();
     }
 
-    /**
-     * Complete Kafka cluster names matching a partial input.
-     *
-     * @param partial the partial cluster name input
-     * @return list of matching cluster names
-     */
     private List<String> completeClusterName(final String partial) {
-        List<String> names = getCachedOrFetch("kafka", () -> {
-            List<Kafka> kafkas = kubernetesClient.resources(Kafka.class)
+        List<String> names = completionCache.getOrFetch("kafka", () ->
+            kubernetesClient.resources(Kafka.class)
                 .inAnyNamespace()
                 .list()
-                .getItems();
-            return kafkas.stream()
+                .getItems()
+                .stream()
                 .map(k -> k.getMetadata().getName())
                 .distinct()
-                .toList();
-        });
+                .toList()
+        );
         return completionHelper.filterByPrefix(names, partial);
     }
 
-    /**
-     * Complete KafkaNodePool names matching a partial input.
-     *
-     * @param partial the partial node pool name input
-     * @return list of matching node pool names
-     */
     private List<String> completeNodePoolName(final String partial) {
-        List<String> names = getCachedOrFetch("nodepool", () -> {
-            List<KafkaNodePool> pools = kubernetesClient.resources(KafkaNodePool.class)
+        List<String> names = completionCache.getOrFetch("nodepool", () ->
+            kubernetesClient.resources(KafkaNodePool.class)
                 .inAnyNamespace()
                 .list()
-                .getItems();
-            return pools.stream()
+                .getItems()
+                .stream()
                 .map(p -> p.getMetadata().getName())
                 .distinct()
-                .toList();
-        });
+                .toList()
+        );
         return completionHelper.filterByPrefix(names, partial);
     }
 
-    /**
-     * Complete KafkaTopic names matching a partial input.
-     *
-     * @param partial the partial topic name input
-     * @return list of matching topic names
-     */
     private List<String> completeTopicName(final String partial) {
-        List<String> names = getCachedOrFetch("topic", () -> {
-            List<KafkaTopic> topics = kubernetesClient.resources(KafkaTopic.class)
+        List<String> names = completionCache.getOrFetch("topic", () ->
+            kubernetesClient.resources(KafkaTopic.class)
                 .inAnyNamespace()
                 .list()
-                .getItems();
-            return topics.stream()
+                .getItems()
+                .stream()
                 .map(t -> t.getMetadata().getName())
                 .distinct()
-                .toList();
-        });
+                .toList()
+        );
         return completionHelper.filterByPrefix(names, partial);
     }
 
-    /**
-     * Complete Strimzi operator deployment names matching a partial input.
-     *
-     * @param partial the partial operator name input
-     * @return list of matching operator deployment names
-     */
     private List<String> completeOperatorName(final String partial) {
-        List<String> names = getCachedOrFetch("operator", () ->
+        List<String> names = completionCache.getOrFetch("operator", () ->
             kubernetesClient.apps().deployments()
                 .inAnyNamespace()
                 .withLabel("app", StrimziConstants.Operator.APP_LABEL_VALUE)
@@ -233,22 +197,5 @@ public class CompletionService {
                 .toList()
         );
         return completionHelper.filterByPrefix(names, partial);
-    }
-
-    private List<String> getCachedOrFetch(final String cacheKey,
-                                           final java.util.function.Supplier<List<String>> fetcher) {
-        CachedNames cached = nameCache.get(cacheKey);
-        if (cached != null && !cached.isExpired()) {
-            return cached.names();
-        }
-        try {
-            List<String> names = fetcher.get();
-            nameCache.put(cacheKey, new CachedNames(names,
-                Instant.now().plusSeconds(cacheTtlSeconds)));
-            return names;
-        } catch (Exception e) {
-            LOG.debugf("Error fetching %s names for completion: %s", cacheKey, e.getMessage());
-            return List.of();
-        }
     }
 }
