+++ strimzi-mcp/src/test/java/io/streamshub/mcp/strimzi/service/KafkaClusterDiagnosticServiceTest.java
@@ -22,8 +22,8 @@ import io.fabric8.kubernetes.client.dsl.V1APIGroupDSL;
 import io.quarkiverse.mcp.server.ToolCallException;
 import io.quarkus.test.InjectMock;
 import io.quarkus.test.junit.QuarkusTest;
+import io.streamshub.mcp.common.util.NamespaceElicitationHelper;
 import io.streamshub.mcp.strimzi.dto.KafkaClusterDiagnosticReport;
-import io.streamshub.mcp.strimzi.util.NamespaceElicitationHelper;
 import io.strimzi.api.kafka.model.kafka.Kafka;
 import io.strimzi.api.kafka.model.kafka.KafkaBuilder;
 import io.strimzi.api.kafka.model.kafka.KafkaStatusBuilder;
