+++ strimzi-mcp/src/test/java/io/streamshub/mcp/strimzi/service/KafkaConnectivityDiagnosticServiceTest.java
@@ -18,8 +18,8 @@ import io.fabric8.kubernetes.client.dsl.RollableScalableResource;
 import io.quarkiverse.mcp.server.ToolCallException;
 import io.quarkus.test.InjectMock;
 import io.quarkus.test.junit.QuarkusTest;
+import io.streamshub.mcp.common.util.NamespaceElicitationHelper;
 import io.streamshub.mcp.strimzi.dto.KafkaConnectivityDiagnosticReport;
-import io.streamshub.mcp.strimzi.util.NamespaceElicitationHelper;
 import io.strimzi.api.kafka.model.kafka.Kafka;
 import io.strimzi.api.kafka.model.kafka.KafkaBuilder;
 import io.strimzi.api.kafka.model.kafka.listener.ListenerStatusBuilder;
